Add persistent state and UI enhancements
Updated the application with persistent state using browser storage. This implementation resulted to User Interface refinements, particularly in managing and displaying the pagination of URLs. Also, "Storage" was added to the permissions list in manifest.json, to allow saving the state in the browser's storage. Lastly, code was improved to increase readability.